refactor(build): code to determine the remote build branch
Removed the reduntant code related to pushing the tagged images
that was spread across multiple files. Created a single `push`
file that is parameterized to work for pushing different containers.

`push` - replaces image-tag and push-* files.

Refactored the cstor-container and cstor-istgt for the following:
- determine the build branch at the beginning and avoid multiple
  if/else blocks
- validate that remote branch exists via git command instead of depending
  on git checkout error

Also, in the process, fixed an issue where tag was always getting created
on master branch instead of the development branch (non-master dot releases).

Signed-off-by: kmova <kiran.mova@openebs.io>